react-travel-app/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow>=9.0.0
openai>=1.0.0
geopy==2.4.0
exifread==3.0.0
unidecode==1.3.7
python-dotenv==1.0.0
aiofiles==23.2.1
PyJWT==2.8.1